generators cli
- `generate-app`
- `generate-admin`
- `generate-module`
- `generate-sql`
- `Installer::register` is deprecated
- Versioning is now my module
- Updating REST permissions to be more predictable and replicable
- Moved file upload capabilities to `Utility/File`
- Added more schemas, relative to something more practical

~20h

TODO:
- Post Image validator
- Permanently add on admin as an app
- Test REST
- More testing on admin
- Test test suites
- Setup kitchen sink on travis
- Move CLI to the `cradlephp/sink-faucet` project

+++ module/utility/src/cli/sql/flush.php
@@ -24,6 +24,11 @@ return function ($request, $response) {
     //truncate all tables
     $tables = $database->getTables();
     foreach ($tables as $table) {
+        //if we just want to flush one table
+        if($request->hasStage('module') && strpos($table, $request->getStage('module')) !== 0) {
+            continue;
+        }
+
         $database->query('TRUNCATE TABLE `' . $table . '`;');
     }
 };
